requirements.txt
Pillow==2.6.0
argparse==1.2.1
reportlab==3.1.8
requests==2.4.3
wsgiref==0.1.2
